feat: Initial commit of HELLA WIP code
Ok, 'HELLA WIP' is an exaggeration but this commit really is mostly to
create a 'before' to compare my 'after' to once I have this ready to
turn in.

This commit implements most of the core features of the memory game.
- The cards are clickable
- The game can be won
- A new game can be started after a game is won
- There is a title card that starts the first game when clicked
- There is a working turns-taken counter that only increments if you
  click a card that creates a non-matching pair of face-up cards
- The positions of the cards are randomized each game
- There are some neat transitions to simulate the card flips
- Basically everything has a glowing shadow to add to the spook factor

This commit does not implement some of the core features required.
- There is no timer ticking up as you play
- There is no star rating when you finish a game

Honestly, it's not in bad shape right now, it's just HOLY HELL the code
needs cleaning up in a major way.