feat: Updated following review feedback
This commit adds some updates requested in my review feedback, whether
I think they're ridiculous or not.

Additional comments were added to the code to explain what functions
like 'buildNewGame' and 'buildCardRow' do.

The readme was updated to be as informative as I could make it,
considering how few actual dependencies this game has (it uses a Google
Font but can work fine without it) and how simple it is to get up and
running (you clone the repo).

A reset button was added below the game board so you can reset the game
at any time, despite the fact that there's no loss condition in the
game. The reset function available on the 'You win' pane remains in
place.

A rating was added to the metrics at the top of the page, despite the
fact that you don't normally give the player a rating for a game that's
in progress. The rating available on the 'You win' pane remains in
place.

There was a significant refactor to break up the large 'handleCardFlip'
function into smaller functions within the conditional logic. With this
change, every function in Memento.js is less than a page long.

Various parts of the process wherein the game is built were
restructured or shuffled around to accommodate the implementation of the
new reset button and the new live rating.